Lift `expr_diverges` to `clippy_utils` as `is_never_expr`
